% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_module.R
\name{histServer}
\alias{histServer}
\title{Histogram Module Server}
\usage{
histServer(id, col_name, filtered_data)
}
\arguments{
\item{id}{Module ID}

\item{col_name}{Column name to plot}

\item{filtered_data}{Reactive expression containing filtered data}
}
\value{
Module server function
}
\description{
Histogram Module Server
}
