% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_module.R
\name{histUI}
\alias{histUI}
\title{Histogram Module UI}
\usage{
histUI(id, col_name)
}
\arguments{
\item{id}{Module ID}

\item{col_name}{Column name to display in the histogram}
}
\value{
A bslib card containing histogram UI elements
}
\description{
Histogram Module UI
}
