% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Palmer Penguins Shiny App}
\usage{
run_app(...)
}
\arguments{
\item{...}{Additional arguments to pass to shiny::runApp}
}
\value{
Runs the Shiny application
}
\description{
Run the Palmer Penguins Shiny App
}
